Moved types to a single module
